CORE-7086-Actualizar-endpoints-carga-masiva
Se ha modificado las URL de las peticiones para que vayan al controlador MassiveResourceApi en vez del ResourceApi.
